Fixed vars
Changed the order that functions executed to get vars working. As a consequence, I had to include "taxonomy" in a lot of the code, which is created by elastic_tab() and was causing problems.

Details on changes to vars are here: 
Details on changes to elastic_tab() related issues are here: https://github.com/AldoCompagnoni/popler/issues/16#issuecomment-285934085